Move hide_mod() to init to reduce binaries size
